Add smoothing of output objects to the detector.
Amazingly enough, it is sufficient to simply add the smoothing
parameter to the settings maps for it to work. Indeed, the smoothing
is handled by the threshold detector, that for ilastik, operates
on the probability map generated by the ilastik command. So there
is only very little to do to make it work here.
Unexpected but pleasant.